pyproject.toml
[project]
name = "lrc-research-prototype"
version = "0.1.0"
description = "Autonomous Box robot agent and Jarvis personal automation system — LRC research prototype"
requires-python = ">=3.9"

dependencies = [
    "anthropic>=0.97.0",
    "pyserial>=3.5",
    "opencv-python>=4.13.0",
    "SpeechRecognition>=3.10.0",
    "pyaudio>=0.2.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
